Removed "Add Another Language" button
Now that there is no home page, there is nowhere for the organization to
return to from the UpdateAccountPage. The same functionality works with
filtering out pages without translatable fields.

In addition, entered data is now pushed to Firestore when the Finished
page is reached. This will prevent users from clicking away from that
page without hitting Done and accidentally losing their input.